# src/docs/openapi.yaml
openapi: 3.0.0
servers:
  - url: 'https://bankapp.com/v1'
    description: Production server
  - url: 'http://localhost:3000/v1'
    description: Testing local server
info:
  version: 1.0.0
  title: Web Banking Application API
  description: The API for the web banking application
paths:
  /register:
    get:
      summary: Serves register page
      tags:
        - auth
      responses:
        '200':
          description: Register page returned successfully
          content:
            text/html:
              schema:
                type: string
                example: <html>...</html>
    post:
      summary: Sending sign-up form to register
      tags:
        - auth
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad Request – Missing or malformed registration data
        '409':
          description: Conflict – Email already in use
...
  /verify-passcode:
    post:
      summary: Verifies the passcode entered by the user
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: The phone number of the user (must be registered)
                passcode:
                  type: string
                  description: The passcode entered by the user (received via SMS)
      responses:
        '200':
          description: Passcode verified successfully, registration completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad Request – Invalid passcode or phone number mismatch
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid passcode or phone number mismatch
        '500':
          description: Internal server error – Error occurred during passcode verification
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Transaction:
      type: object
      properties:
        sender:
          $ref: '#/components/schemas/User'
        receiver:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        email:
          type: string
        id:
          type: integer
        phone:
          type: string
        amount:
          type: integer
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    NewUser:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        phone:
          type: string
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT for authenticated requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...